@@ -1,5 +1,11 @@
 <?php
 
+use App\Http\Controllers\ArticleController;
+use App\Http\Controllers\Auth\LoginController;
+use App\Http\Controllers\CommentController;
+use App\Http\Controllers\ContactsController;
+use App\Http\Controllers\IndexController;
+use App\Http\Controllers\PortfolioController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -17,43 +23,43 @@ Route::get('/', function () {
     return view('welcome');
 });
 
-Route::resource('/', 'IndexController', [
+Route::resource('/', IndexController::class, [
     'only' => 'index',
     'names' => ['index' => 'home'],
 ]);
 
-Route::resource('portfolios', 'PortfolioController', [
+Route::resource('portfolios', PortfolioController::class, [
     'parameters' => [
         'portfolios' => 'alias'
     ],
 ]);
 
-Route::resource('articles', 'ArticleController', [
+Route::resource('articles', ArticleController::class, [
     'parameters' => [
         'articles' => 'alias'
     ],
 ]);
 
 Route::get('articles/cat/{alias?}', [
-    'uses' => 'ArticleController@index',
+    'uses' => [ArticleController::class, 'index'],
     'as' => 'articlesCat'
 ])->where('alias','[\w-]+');
 
-Route::resource('comment', 'CommentController', ['only' => ['store']]);
+Route::resource('comment', CommentController::class, ['only' => ['store']]);
 
-Route::match(['get','post'],'/contacts',['uses'=>'ContactsController@index','as'=>'contacts']);
+Route::match(['get','post'],'/contacts',['uses' => 'App\Http\Controllers\ContactsController@index','as'=>'contacts']);
 
-Route::get('login', 'Auth\LoginController@showLoginForm')->name('login');
-Route::post('login', 'Auth\LoginController@login');
-Route::get('logout', 'Auth\LoginController@logout')->name('admin.logout');
+Route::get('login', [LoginController::class, 'showLoginForm'])->name('login');
+Route::post('login', [LoginController::class, 'login']);
+Route::get('logout', [LoginController::class, 'logout'])->name('admin.logout');
 
 
 // Admin section
 Route::group(['prefix' => 'admin', 'middleware' => 'auth'], function() {
 
-    Route::get('/', ['uses' => 'Admin\IndexController@index', 'as' => 'adminIndex']);
-    Route::resource('/articles', 'Admin\ArticlesController', ['names' => 'admin.articles']);
-    Route::resource('/permissions', 'Admin\PermissionsController', ['names' => 'admin.permissions']);
-    Route::resource('/menus', 'Admin\MenusController', ['names' => 'admin.menus']);
+    Route::get('/', ['uses' => 'App\Http\Controllers\Admin\IndexController@index', 'as' => 'adminIndex']);
+    Route::resource('/articles', 'App\Http\Controllers\Admin\ArticlesController', ['names' => 'admin.articles']);
+    Route::resource('/permissions', 'App\Http\Controllers\Admin\PermissionsController', ['names' => 'admin.permissions']);
+    Route::resource('/menus', 'App\Http\Controllers\Admin\MenusController', ['names' => 'admin.menus']);
 
 });
