@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use App\Article;
-use App\Comment;
+use App\Models\Article;
+use App\Models\Comment;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
