@@ -1,7 +1,10 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
+use App\Models\Category;
+use App\Models\Comment;
+use App\Models\User;
 use Illuminate\Database\Eloquent\Model;
 
 class Article extends Model
