@@ -1,7 +1,8 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
+use App\Models\Article;
 use Illuminate\Database\Eloquent\Model;
 
 class Category extends Model
