@@ -1,7 +1,9 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
+use App\Models\User;
+use App\Models\Permission;
 use Illuminate\Database\Eloquent\Model;
 
 class Role extends Model
