@@ -2,8 +2,8 @@
 
 namespace App\Policies;
 
-use App\Article;
-use App\User;
+use App\Models\Article;
+use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
 class ArticlePolicy
