@@ -4,7 +4,7 @@ namespace App\Policies;
 
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-use App\User;
+use App\Models\User;
 
 class PermissionPolicy
 {
