@@ -3,7 +3,7 @@
 
 namespace App\Repositories;
 
-use App\Comment;
+use App\Models\Comment;
 
 
 class CommentsRepository extends Repository
