@@ -2,12 +2,12 @@
 
 namespace App\Repositories;
 
-use App\Role;
+use App\Models\Role;
 
 class RolesRepository extends Repository {
 
 	public function __construct(Role $role) {
 		$this->model = $role;
 	}
-	
+
 }
