@@ -3,7 +3,7 @@
 
 namespace App\Repositories;
 
-use App\Slider;
+use App\Models\Slider;
 
 
 class SlidersRepository extends Repository
